# internal/ginapi/docs/swagger.yaml
basePath: /api/v1
definitions:
  ginapi.BuildingBody:
    properties:
      city:
        type: string
      floors_count:
        type: integer
      handover_year:
        type: integer
      name:
        type: string
    required:
    - city
    - floors_count
    - handover_year
    - name
    type: object
  ginapi.BuildingView:
    properties:
      city:
        type: string
      floors_count:
        type: integer
      handover_year:
        type: integer
      id:
        type: integer
      name:
        type: string
    type: object
  ginapi.Error:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
host: localhost:8000
info:
  contact:
    email: rylenko@tuta.io
    name: rylenko
    url: http://github.com/rylenko
  description: Swagger API of Leadgen Market Task
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Leadgen Market Task
  version: "1.0"
paths:
  /buildings:
    get:
      consumes:
      - application/json
      description: Gets all buildings according to passed filter paramters
      operationId: getall-buildings
      parameters:
      - description: city filter
        in: query
        name: city
        type: string
      - description: handover year filter
        in: query
        name: handover_year
        type: integer
      - description: floors count filter
        in: query
        name: floors_count
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ginapi.BuildingView'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ginapi.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ginapi.Error'
      summary: Gets all buildings
      tags:
      - building
    post:
      consumes:
      - application/json
      description: Creates a new building using passed data
      operationId: create-building
      parameters:
      - description: Create building
        in: body
        name: building
        required: true
        schema:
          $ref: '#/definitions/ginapi.BuildingBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ginapi.BuildingView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ginapi.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ginapi.Error'
      summary: Creates a new building
      tags:
      - building
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"